Added all config values to config file.

+++ src/org/wildstang/yearly/auto/programs/ShootGear.java
@@ -23,7 +23,7 @@ public class ShootGear extends AutoProgram
       super.initialize();
       
       // Read config values
-      delayWhileShooting = Core.getConfigManager().getConfig().getDouble(this.getClass().getName() + ".delayWhileShooting", 2000);
+      delayWhileShooting = Core.getConfigManager().getConfig().getInt(this.getClass().getName() + ".delayWhileShooting", 2000);
    }
    protected void defineSteps()
    {
